style: define traits in `pool.rs` mod
